+++ darglint/errors.py
@@ -29,18 +29,18 @@ class DarglintError(BaseException):
     # The shortest error message possible.  Should use abbreviated
     # symbols.  Can be combined with the general message to give
     # a more in-depth message.
-    terse_message = None
+    terse_message = None  # type: str
 
     # General messages should be general in nature. (They are used
     # at the lowest verbosity setting.)  They should be about
     # the nature of the error, and not about this particular instance.
     # Should not end in any punctuation (and should allow lists of
     # instances -- the terse message -- after it.)
-    general_message = None
+    general_message = None  # type: str
 
     # A unique human-readable identifier for this type of error.
     # See the description of error code groups above.
-    error_code = None
+    error_code = None  # type: str
 
     def message(self, verbosity: int=1) -> str:
         """Get the message for this error, according to the verbosity.
@@ -68,7 +68,7 @@ class DarglintError(BaseException):
             raise Exception('Unrecognized verbosity setting, {}.'.format(
                 verbosity))
 
-    def __init__(self, function: ast.FunctionDef):
+    def __init__(self, function: ast.FunctionDef) -> None:
         """Create a new exception with a message and line number.
 
         Raises:
@@ -94,7 +94,7 @@ class GenericSyntaxError(DarglintError):
 
     error_code = 'S001'
 
-    def __init__(self, function: ast.FunctionDef, message: str):
+    def __init__(self, function: ast.FunctionDef, message: str) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -113,7 +113,7 @@ class MissingParameterError(DarglintError):
 
     error_code = 'I101'
 
-    def __init__(self, function: ast.FunctionDef, name: str):
+    def __init__(self, function: ast.FunctionDef, name: str) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -131,7 +131,7 @@ class ExcessParameterError(DarglintError):
 
     error_code = 'I102'
 
-    def __init__(self, function: ast.FunctionDef, name: str):
+    def __init__(self, function: ast.FunctionDef, name: str) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -153,7 +153,7 @@ class ParameterTypeMismatchError(DarglintError):
                  function: ast.FunctionDef,
                  name: str,
                  expected: str,
-                 actual: str):
+                 actual: str) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -180,7 +180,7 @@ class MissingReturnError(DarglintError):
 
     error_code = 'I201'
 
-    def __init__(self, function: ast.FunctionDef):
+    def __init__(self, function: ast.FunctionDef) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -198,7 +198,7 @@ class ExcessReturnError(DarglintError):
 
     error_code = 'I202'
 
-    def __init__(self, function: ast.FunctionDef):
+    def __init__(self, function: ast.FunctionDef) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -219,7 +219,7 @@ class ReturnTypeMismatchError(DarglintError):
     def __init__(self,
                  function: ast.FunctionDef,
                  expected: str,
-                 actual: str):
+                 actual: str) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -243,7 +243,7 @@ class MissingYieldError(DarglintError):
 
     error_code = 'I301'
 
-    def __init__(self, function: ast.FunctionDef):
+    def __init__(self, function: ast.FunctionDef) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -261,7 +261,7 @@ class ExcessYieldError(DarglintError):
 
     error_code = 'I302'
 
-    def __init__(self, function: ast.FunctionDef):
+    def __init__(self, function: ast.FunctionDef) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -279,7 +279,7 @@ class MissingRaiseError(DarglintError):
 
     error_code = 'I401'
 
-    def __init__(self, function: ast.FunctionDef, name: str):
+    def __init__(self, function: ast.FunctionDef, name: str) -> None:
         """Instantiate the error's message.
 
         Args:
@@ -305,7 +305,7 @@ class ExcessRaiseError(DarglintError):
 
     error_code = 'I402'
 
-    def __init__(self, function: ast.FunctionDef, name: str):
+    def __init__(self, function: ast.FunctionDef, name: str) -> None:
         """Instantiate the error's message.
 
         Args: